+++ changelog.d/20231120_134843_klakhov_improve_invite_users.md
@@ -1,4 +1,5 @@
-### Changed
+### Added
 
-- Inviting all users to organization by email
+- Implemented a feature that allows to invite users to organization using email
+  or internal CVAT invitation
   (<https://github.com/opencv/cvat/pull/7138>)
