+++ cvat-core/src/api-implementation.ts
@@ -355,13 +355,7 @@ export default function implementAPI(cvat) {
         return orgSlug;
     };
 
-    cvat.organizations.rejectInvitation.implementation = async (
-        key,
-    ) => {
-        await serverProxy.organizations.rejectInvitation(
-            key,
-        );
-    };
+    cvat.organizations.rejectInvitation.implementation = serverProxy.organizations.rejectInvitation;
 
     cvat.organizations.invitations.implementation = async (filter: InvitationsFilter = {}) => {
         const invitationsData = await serverProxy.organizations.invitations(filter);
