+++ cvat-ui/package.json
@@ -1,6 +1,6 @@
 {
   "name": "cvat-ui",
-  "version": "1.59.1",
+  "version": "1.60.0",
   "description": "CVAT single-page application",
   "main": "src/index.tsx",
   "scripts": {
